requirements.txt
Flask==2.0.1
pandas==1.3.3
pymongo[srv]==3.12.0
Werkzeug==2.2.2
openpyxl
python-dotenv

